+++ src/test/java/com/epam/aidial/cfg/service/config/export/CoreConfigReloadServiceTest.java
@@ -1,8 +1,6 @@
-package com.epam.aidial.cfg.service.export;
+package com.epam.aidial.cfg.service.config.export;
 
 import com.epam.aidial.cfg.client.CoreConfigClient;
-import com.epam.aidial.cfg.service.config.export.ConfigExportFacade;
-import com.epam.aidial.cfg.service.config.export.CoreConfigReloadService;
 import feign.FeignException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
