+++ src/test/java/com/epam/aidial/cfg/service/config/impl/storage/ConfigSplitterTest.java
@@ -1,8 +1,6 @@
-package com.epam.aidial.cfg.service.impl.storage;
+package com.epam.aidial.cfg.service.config.impl.storage;
 
 import com.epam.aidial.cfg.configuration.JsonMapperConfiguration;
-import com.epam.aidial.cfg.service.config.impl.storage.ConfigPart;
-import com.epam.aidial.cfg.service.config.impl.storage.ConfigSplitter;
 import com.epam.aidial.core.config.Assistants;
 import com.epam.aidial.core.config.Config;
 import com.epam.aidial.core.config.CoreAddon;
