+++ src/test/java/com/epam/aidial/cfg/service/config/impl/storage/GcpVaultConfigSourceTest.java
@@ -1,17 +1,18 @@
-package com.epam.aidial.cfg.service.impl.storage;
+package com.epam.aidial.cfg.service.config.impl.storage;
 
-import com.azure.core.exception.ResourceNotFoundException;
-import com.azure.security.keyvault.secrets.SecretClient;
-import com.azure.security.keyvault.secrets.models.KeyVaultSecret;
-import com.epam.aidial.cfg.service.config.impl.storage.AzureKeyVaultConfigSource;
-import com.epam.aidial.cfg.service.config.impl.storage.ConfigMerger;
-import com.epam.aidial.cfg.service.config.impl.storage.ConfigPart;
-import com.epam.aidial.cfg.service.config.impl.storage.ConfigSplitter;
 import com.epam.aidial.cfg.service.config.transfer.VersionAwareFieldFilter;
 import com.epam.aidial.core.config.Config;
 import com.epam.aidial.core.config.CoreKey;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.api.gax.httpjson.HttpJsonStatusCode;
+import com.google.api.gax.rpc.NotFoundException;
+import com.google.api.gax.rpc.StatusCode;
+import com.google.cloud.secretmanager.v1.AccessSecretVersionResponse;
+import com.google.cloud.secretmanager.v1.SecretManagerServiceClient;
+import com.google.cloud.secretmanager.v1.SecretPayload;
+import com.google.cloud.secretmanager.v1.SecretVersionName;
+import com.google.protobuf.ByteString;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -19,8 +20,10 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import java.io.UnsupportedEncodingException;
 import java.util.List;
 import java.util.Map;
 
@@ -32,16 +35,14 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-class AzureKeyVaultConfigSourceTest {
+class GcpVaultConfigSourceTest {
 
     @Mock
-    private SecretClient secretClient;
+    private SecretManagerServiceClient secretClient;
     @Captor
-    private ArgumentCaptor<KeyVaultSecret> keyVaultSecretArgumentCaptor;
-    private String expirationTimeUnit = "MONTHS";
-    private long expirationPeriod = 3;
+    private ArgumentCaptor<SecretPayload> keyVaultSecretArgumentCaptor;
     private ObjectMapper objectMapper = new ObjectMapper();
-    private AzureKeyVaultConfigSource source;
+    private GcpVaultConfigSource source;
     @Mock
     private ConfigSplitter configSplitter;
     @Mock
@@ -50,9 +51,8 @@ class AzureKeyVaultConfigSourceTest {
     private VersionAwareFieldFilter versionAwareFieldFilter;
 
     @Test
-    void writeConfig_shouldWriteSecret() throws JsonProcessingException {
-        source = new AzureKeyVaultConfigSource(versionAwareFieldFilter, configSplitter, configMerger, List.of("secret1"),
-            secretClient, expirationTimeUnit, expirationPeriod, objectMapper);
+    void writeConfig_shouldWriteSecret() throws JsonProcessingException, UnsupportedEncodingException {
+        source = new GcpVaultConfigSource(versionAwareFieldFilter, secretClient, configSplitter, configMerger, List.of("secret1"), objectMapper, "projectId");
 
         Config config = new Config();
         CoreKey key = new CoreKey();
@@ -62,25 +62,26 @@ class AzureKeyVaultConfigSourceTest {
 
         when(versionAwareFieldFilter.filterForTargetVersion(config)).thenReturn(config);
 
-        when(secretClient.getSecret(any(String.class))).thenReturn(new KeyVaultSecret("secret", "{}"));
+        ByteString body = ByteString.copyFrom("{}", "utf-8");
+        SecretPayload payload = SecretPayload.newBuilder().setData(body).build();
+        AccessSecretVersionResponse response = AccessSecretVersionResponse.newBuilder().setPayload(payload).build();
+        when(secretClient.accessSecretVersion(any(SecretVersionName.class))).thenReturn(response);
         Config secretConfig = new Config();
         secretConfig.setKeys(Map.of("key1", key));
 
         source.writeConfig(config, false);
 
         verify(configSplitter, times(0)).splitConfig(any(), any(), anyInt(), eq(1));
-        verify(secretClient, times(1)).setSecret(keyVaultSecretArgumentCaptor.capture());
+        verify(secretClient, times(1)).addSecretVersion(Mockito.eq("secret1"), keyVaultSecretArgumentCaptor.capture());
 
-        List<KeyVaultSecret> allValues = keyVaultSecretArgumentCaptor.getAllValues();
+        List<SecretPayload> allValues = keyVaultSecretArgumentCaptor.getAllValues();
         Assertions.assertEquals(allValues.size(), 1);
-        Assertions.assertEquals(allValues.get(0).getName(), "secret1");
-        Assertions.assertEquals(objectMapper.readValue(allValues.get(0).getValue(), Config.class), config);
+        Assertions.assertEquals(objectMapper.readValue(allValues.get(0).getData().toStringUtf8(), Config.class), config);
     }
 
     @Test
-    void writeConfig_shouldWriteWithoutSplit() throws JsonProcessingException {
-        source = new AzureKeyVaultConfigSource(versionAwareFieldFilter, configSplitter, configMerger, List.of("secret1", "secret2"),
-            secretClient, expirationTimeUnit, expirationPeriod, objectMapper);
+    void writeConfig_shouldWriteWithoutSplit() throws JsonProcessingException, UnsupportedEncodingException {
+        source = new GcpVaultConfigSource(versionAwareFieldFilter, secretClient, configSplitter, configMerger, List.of("secret1", "secret2"), objectMapper, "projectId");
 
         Config config = new Config();
         CoreKey key = new CoreKey();
@@ -90,7 +91,10 @@ class AzureKeyVaultConfigSourceTest {
 
         when(versionAwareFieldFilter.filterForTargetVersion(config)).thenReturn(config);
 
-        when(secretClient.getSecret(any(String.class))).thenReturn(new KeyVaultSecret("secret", "{}"));
+        ByteString body = ByteString.copyFrom("{}", "utf-8");
+        SecretPayload payload = SecretPayload.newBuilder().setData(body).build();
+        AccessSecretVersionResponse response = AccessSecretVersionResponse.newBuilder().setPayload(payload).build();
+        when(secretClient.accessSecretVersion(any(SecretVersionName.class))).thenReturn(response);
         Config secretConfig = new Config();
         secretConfig.setKeys(Map.of("key1", key));
         ConfigPart configPart = new ConfigPart(secretConfig, objectMapper.writeValueAsString(secretConfig));
@@ -98,20 +102,18 @@ class AzureKeyVaultConfigSourceTest {
 
         source.writeConfig(config, false);
 
-        verify(secretClient, times(2)).setSecret(keyVaultSecretArgumentCaptor.capture());
+        verify(secretClient, times(1)).addSecretVersion(Mockito.eq("secret1"), keyVaultSecretArgumentCaptor.capture());
+        verify(secretClient, times(1)).addSecretVersion(Mockito.eq("secret2"), keyVaultSecretArgumentCaptor.capture());
 
-        List<KeyVaultSecret> allValues = keyVaultSecretArgumentCaptor.getAllValues();
+        List<SecretPayload> allValues = keyVaultSecretArgumentCaptor.getAllValues();
         Assertions.assertEquals(allValues.size(), 2);
-        Assertions.assertEquals(allValues.get(0).getName(), "secret1");
-        Assertions.assertEquals(objectMapper.readValue(allValues.get(0).getValue(), Config.class), config);
-        Assertions.assertEquals(allValues.get(1).getName(), "secret2");
-        Assertions.assertEquals(objectMapper.readValue(allValues.get(1).getValue(), Config.class), new Config());
+        Assertions.assertEquals(objectMapper.readValue(allValues.get(0).getData().toStringUtf8(), Config.class), config);
+        Assertions.assertEquals(objectMapper.readValue(allValues.get(1).getData().toStringUtf8(), Config.class), new Config());
     }
 
     @Test
     void writeConfig_shouldThrowExceptionNotEnoughSpace() throws JsonProcessingException {
-        source = new AzureKeyVaultConfigSource(versionAwareFieldFilter, configSplitter, configMerger, List.of("secret1", "secret2"),
-            secretClient, expirationTimeUnit, expirationPeriod, objectMapper);
+        source = new GcpVaultConfigSource(versionAwareFieldFilter, secretClient, configSplitter, configMerger, List.of("secret1", "secret2"), objectMapper, "projectId");
 
         Config config = new Config();
         CoreKey key = new CoreKey();
@@ -132,8 +134,7 @@ class AzureKeyVaultConfigSourceTest {
 
     @Test
     void writeConfig_shouldCreateResourceWhenTrue() throws JsonProcessingException {
-        source = new AzureKeyVaultConfigSource(versionAwareFieldFilter, configSplitter, configMerger, List.of("secret1"),
-            secretClient, expirationTimeUnit, expirationPeriod, objectMapper);
+        source = new GcpVaultConfigSource(versionAwareFieldFilter, secretClient, configSplitter, configMerger, List.of("secret1"), objectMapper, "projectId");
 
         Config config = new Config();
         CoreKey key = new CoreKey();
@@ -143,23 +144,21 @@ class AzureKeyVaultConfigSourceTest {
 
         when(versionAwareFieldFilter.filterForTargetVersion(config)).thenReturn(config);
 
-        when(secretClient.getSecret(any(String.class)))
-                .thenThrow(new ResourceNotFoundException("Secret not found", null));
+        when(secretClient.accessSecretVersion(any(SecretVersionName.class)))
+                .thenThrow(new NotFoundException(new RuntimeException(), HttpJsonStatusCode.of(StatusCode.Code.NOT_FOUND), false));
 
         source.writeConfig(config, true);
 
-        verify(secretClient, times(1)).setSecret(keyVaultSecretArgumentCaptor.capture());
+        verify(secretClient, times(1)).addSecretVersion(Mockito.eq("secret1"), keyVaultSecretArgumentCaptor.capture());
 
-        List<KeyVaultSecret> allValues = keyVaultSecretArgumentCaptor.getAllValues();
+        List<SecretPayload> allValues = keyVaultSecretArgumentCaptor.getAllValues();
         Assertions.assertEquals(1, allValues.size());
-        Assertions.assertEquals("secret1", allValues.get(0).getName());
-        Assertions.assertEquals(objectMapper.readValue(allValues.get(0).getValue(), Config.class), config);
+        Assertions.assertEquals(config, objectMapper.readValue(allValues.get(0).getData().toStringUtf8(), Config.class));
     }
 
     @Test
     void writeConfig_shouldThrowExceptionWhenResourceNotFoundAndCreateResourcesFalse() {
-        source = new AzureKeyVaultConfigSource(versionAwareFieldFilter, configSplitter, configMerger, List.of("secret1"),
-            secretClient, expirationTimeUnit, expirationPeriod, objectMapper);
+        source = new GcpVaultConfigSource(versionAwareFieldFilter, secretClient, configSplitter, configMerger, List.of("secret1"), objectMapper, "projectId");
 
         Config config = new Config();
         CoreKey key = new CoreKey();
@@ -169,8 +168,8 @@ class AzureKeyVaultConfigSourceTest {
 
         when(versionAwareFieldFilter.filterForTargetVersion(config)).thenReturn(config);
 
-        when(secretClient.getSecret(any(String.class)))
-                .thenThrow(new ResourceNotFoundException("Secret not found", null));
+        when(secretClient.accessSecretVersion(any(SecretVersionName.class)))
+                .thenThrow(new NotFoundException(new RuntimeException(), HttpJsonStatusCode.of(StatusCode.Code.NOT_FOUND), false));
 
         IllegalStateException exception = Assertions.assertThrows(
                 IllegalStateException.class,
@@ -185,5 +184,4 @@ class AzureKeyVaultConfigSourceTest {
         Config config = new Config();
         return new ConfigPart(config, objectMapper.writeValueAsString(config));
     }
-
 }
